Fix duplicate entries in catalog

+++ build/travis/catalog/s3.js
@@ -6,8 +6,9 @@ import promisify from 'util';
 const S3_API_VERSION = '2006-03-01';
 
 class S3Store {
-  constructor(bucket, accessKeyId, secretAccessKey) {
+  constructor(bucket, dry, accessKeyId, secretAccessKey) {
     this.bucket = bucket;
+    this.dry = dry;
 
     this.s3 = new AWS.S3({
       accessKeyId: accessKeyId,
@@ -58,14 +59,19 @@ class S3Store {
   }
 
   async writeJson(filename, data) {
-    let jsonData = JSON.stringify(data); //, null, '\t');
-    return this.s3
-      .putObject({
-        Key: filename,
-        Body: jsonData,
-        ContentType: 'application/json'
-      })
-      .promise();
+    if (this.dry) {
+      console.log('Dry run: writeJson(' + filename + '):');
+      console.log(JSON.stringify(data, null, '\t'));
+    } else {
+      let jsonData = JSON.stringify(data); //, null, '\t');
+      return this.s3
+        .putObject({
+          Key: filename,
+          Body: jsonData,
+          ContentType: 'application/json'
+        })
+        .promise();
+    }
   }
 
   async readJson(filename) {
@@ -74,7 +80,11 @@ class S3Store {
   }
 
   async deleteObject(filename) {
-    return this.s3.deleteObject({ Key: filename }).promise();
+    if (this.dry) {
+      console.log('Dry run: deleteObject(' + filename + ')');
+    } else {
+      return this.s3.deleteObject({ Key: filename }).promise();
+    }
   }
 }
